work in progress refactoring
